src/main/java/com/example/study/leetcode/easy/SymmetricTree.java
package com.example.study.leetcode.easy;

import org.springframework.stereotype.Component;

@Component
class SymmetricTree {
    /**
     * 이진 트리의 루트가 주어지면 그것이 자신의 거울인지(즉, 중심을 중심으로 대칭인지) 확인합니다.
     */
    public boolean isSymmetric(TreeNode root) {
        return isMirror(root, root);
    }

    public boolean isMirror(TreeNode t1, TreeNode t2) {
        if (t1 == null && t2 == null) return true;
        if (t1 == null || t2 == null) return false;
        return (t1.val == t2.val)
                && isMirror(t1.right, t2.left)
                && isMirror(t1.left, t2.right);
    }

    public static class TreeNode {
        int val;
        TreeNode left;
        TreeNode right;
        TreeNode() {}
        TreeNode(int val) { this.val = val; }
        TreeNode(int val, TreeNode left, TreeNode right) {
            this.val = val;
            this.left = left;
            this.right = right;
        }
    }
}
